+++ flake.nix
@@ -35,7 +35,7 @@
       with pkgs;
       let
         name = "snapcache";
-        version = "1.2.0";
+        version = "1.2.1";
         composerSrc = pkgs.lib.cleanSourceWith {
           src = self;
           filter =